# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Feature Sequence Store"
  version: "1.0.0"
  title: "Feature Sequence Store"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "hsyoo@anl.gov"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v1"
tags:
- name: "Sequence"
  description: "Feature Sequence"
schemes:
- "http"
paths:
  /feature_sequence:
    post:
      tags:
      - "Sequence"
      summary: "Add a new sequence"
      description: ""
      operationId: "addSequence"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "sequence"
        required: true
        schema:
          $ref: "#/definitions/FeatureSequence"
      responses:
        "405":
          description: "Invalid input"
      security:
      - api_key: []
      x-swagger-router-controller: "Sequence"
  /feature_sequence/batch/:
    post:
      tags:
      - "Sequence"
      summary: "Add new sequences"
      description: ""
      operationId: "addMultipleSequences"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "sequences"
        required: true
        schema:
          $ref: "#/definitions/ArrayOfFeatureSequence"
      responses:
        "405":
          description: "Invalid input"
      security:
      - api_key: []
      x-swagger-router-controller: "Sequence"
  /feature_sequence/{id}:
    get:
      tags:
      - "Sequence"
      summary: "Find sequence by ID"
      description: "Returns a single sequence"
      operationId: "getSequenceById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of sequence"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/FeatureSequence"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Sequence not found"
      security:
      - api_key: []
      x-swagger-router-controller: "Sequence"
  /feature_sequence/batch/{ids}:
    get:
      tags:
      - "Sequence"
      summary: "Find multiple sequence by ID"
      description: "Returns a single sequence"
      operationId: "getMultipleSequencesById"
      produces:
      - "application/json"
      parameters:
      - name: "ids"
        in: "path"
        description: "List of IDs comma seperated"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/FeatureSequence"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Sequence not found"
      security:
      - api_key: []
      x-swagger-router-controller: "Sequence"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  FeatureSequence:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "string"
        description: "md5 hash"
      sequence:
        type: "string"
        example: "ACTG"
      type:
        type: "string"
        description: "AA or NA sequence"
        enum:
        - "AA"
        - "NA"
    example:
      sequence: "ACTG"
      id: "id"
      type: "AA"
  ArrayOfFeatureSequence:
    type: "array"
    items:
      $ref: "#/definitions/FeatureSequence"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"